add sqlite to golang client

+++ client-go/internal/tcpclient/packet.go
@@ -61,7 +61,7 @@ func (p Packet) payload(s *TCPServer) ([]byte, error) {
 
 	message = append(message, p.Data...)
 
-	message = append(utils.IntToBytes(len(message)), message...)
+	message = append(utils.IntToBytes(int64(len(message))), message...)
 	return message, nil
 }
